pyproject.toml
[project]
authors = [
    { name = "Chris Gregory", email = "christopher.b.gregory@gmail.com" },
]
dependencies = [
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "typer>=0.16.0",
]
description = "Zhed puzzle game solver."
keywords = ["zhed", "puzzle", "game", "solver"]
name = "zhed"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[tool.uv]
dev-dependencies = [
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
    "semver>=3.0.4",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project.urls]
repository = "https://github.com/gregorybchris/zhed"

[project.scripts]
zhed = "zhed.cli:app"

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
src = ["src"]

[tool.ruff.lint]
ignore = ["D100", "D104", "FIX002", "TD002", "TD003"]
select = [
    "A",
    "ARG",
    "B",
    "BLE",
    "C4",
    "E",
    "EM",
    "ERA",
    "F",
    "FIX",
    "G",
    "I",
    "ICN",
    "ISC",
    "LOG",
    "N",
    "PL",
    "PT",
    "PTH",
    "PLR",
    "RET",
    "RUF",
    "SIM",
    "SLF",
    "T20",
    "TCH",
    "TD",
    "TID",
    "W",
]

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.extend-per-file-ignores]
"**/*.py" = ["A001", "A002"]
"**/tests/**/*.py" = ["SLF", "PLR2004", "PLR6301"]
"src/zhed/cli.py" = ["T201", "T203"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["functional"]
